refactor: extract layout wrapping logic to proper Next.js App Router patterns
- Create (authenticated) route group with centralized layout
- Move authenticated pages to route group structure
- Remove manual MainLayout wrapping from individual pages
- Centralize auth guards and loading states in layout
- Improve maintainability and follow Next.js best practices

File changes:
- Created src/app/(authenticated)/layout.tsx with auth logic
- Moved dashboard, account, admin, servers to authenticated route group
- Simplified page components to focus on content only
- Removed duplicate auth checks and layout wrapping

Benefits:
- Consistent layout application across authenticated pages
- Better separation of concerns
- Follows Next.js App Router conventions
- Easier to maintain and modify layouts
- Eliminates code duplication

Resolves #6
