+++ src/app/(authenticated)/account/page.tsx
@@ -1,0 +1,5 @@
+import { AccountSettings } from "@/components/account/account-settings";
+
+export default function AccountPage() {
+  return <AccountSettings />;
+}
