+++ src/app/(authenticated)/dashboard/page.tsx
@@ -1,0 +1,5 @@
+import { ServerDashboard } from "@/components/server/server-dashboard";
+
+export default function DashboardPage() {
+  return <ServerDashboard />;
+}
