+++ src/app/(authenticated)/servers/[id]/server-detail.module.css
